# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "internet-date API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/swipes": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Like / pass",
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwipeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwipeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List matches",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/inbox": {
      "get": {
        "tags": [
          "messaging"
        ],
        "summary": "Inbox list",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversationId}": {
      "get": {
        "tags": [
          "messaging"
        ],
        "summary": "Get conversation",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "cursorId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversationId}/messages": {
      "post": {
        "tags": [
          "messaging"
        ],
        "summary": "Send message",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSendBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSendResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{messageId}/read": {
      "post": {
        "tags": [
          "messaging"
        ],
        "summary": "Mark read",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/upload": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Upload media",
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MediaUploadBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{mediaId}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get media",
        "security": [],
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^\\d+$"
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "MediaStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "READY",
          "FAILED"
        ]
      },
...
      "SwipeAction": {
        "type": "string",
        "enum": [
          "LIKE",
          "PASS"
        ]
      },
...
      "MediaUrls": {
        "type": "object",
        "properties": {
          "original": {
            "type": "string"
          },
          "thumb": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "original"
        ]
      },
      "MediaUploadResponse": {
        "type": "object",
        "properties": {
          "mediaId": {
            "$ref": "#/components/schemas/Id"
          },
          "status": {
            "$ref": "#/components/schemas/MediaStatus"
          },
          "mimeType": {
            "type": "string"
          },
          "urls": {
            "$ref": "#/components/schemas/MediaUrls"
          }
        },
        "required": [
          "mediaId",
          "status",
          "mimeType",
          "urls"
        ]
      },
      "MediaResponse": {
        "type": "object",
        "properties": {
          "mediaId": {
            "$ref": "#/components/schemas/Id"
          },
          "status": {
            "$ref": "#/components/schemas/MediaStatus"
          },
          "mimeType": {
            "type": [
              "string",
              "null"
            ]
          },
          "sizeBytes": {
            "type": [
              "number",
              "null"
            ]
          },
          "width": {
            "type": [
              "number",
              "null"
            ]
          },
          "height": {
            "type": [
              "number",
              "null"
            ]
          },
          "durationSec": {
            "type": [
              "number",
              "null"
            ]
          },
          "urls": {
            "$ref": "#/components/schemas/MediaUrls"
          }
        },
        "required": [
          "mediaId",
          "status",
          "urls"
        ]
      },
      "MediaUploadBody": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
...
      "SwipeBody": {
        "type": "object",
        "properties": {
          "toUserId": {
            "$ref": "#/components/schemas/Id"
          },
          "action": {
            "$ref": "#/components/schemas/SwipeAction"
          }
        },
        "required": [
          "toUserId",
          "action"
        ]
      },
      "SwipeResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "matched": {
            "type": "boolean"
          },
          "matchId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "MatchUserProfile": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "locationText": {
            "type": [
              "string",
              "null"
            ]
          },
          "intent": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "MatchUser": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchUserProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "MatchItem": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "userAId": {
            "$ref": "#/components/schemas/Id"
          },
          "userBId": {
            "$ref": "#/components/schemas/Id"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "conversation": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/Id"
                  }
                },
                "required": [
                  "id"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "userA": {
            "$ref": "#/components/schemas/MatchUser"
          },
          "userB": {
            "$ref": "#/components/schemas/MatchUser"
          }
        },
        "required": [
          "id",
          "userAId",
          "userBId",
          "updatedAt",
          "userA",
          "userB"
        ]
      },
      "MatchListResponse": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchItem"
            }
          }
        },
        "required": [
          "matches"
        ]
      },
      "InboxUserProfile": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InboxUser": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InboxUserProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "InboxMessage": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "senderId": {
            "$ref": "#/components/schemas/Id"
          }
        },
        "required": [
          "id",
          "body",
          "createdAt",
          "senderId"
        ]
      },
      "InboxConversation": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "otherUser": {
            "$ref": "#/components/schemas/InboxUser"
          },
          "lastMessage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InboxMessage"
              },
              {
                "type": "null"
              }
            ]
          },
          "unreadCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "updatedAt",
          "otherUser",
          "lastMessage",
          "unreadCount"
        ]
      },
      "InboxResponse": {
        "type": "object",
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InboxConversation"
            }
          }
        },
        "required": [
          "conversations"
        ]
      },
      "MessageItem": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "body": {
            "type": "string"
          },
          "senderId": {
            "$ref": "#/components/schemas/Id"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "body",
          "senderId",
          "createdAt"
        ]
      },
      "MessageListResponse": {
        "type": "object",
        "properties": {
          "conversationId": {
            "$ref": "#/components/schemas/Id"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageItem"
            }
          },
          "nextCursorId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "conversationId",
          "messages",
          "nextCursorId"
        ]
      },
      "MessageSendBody": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "MessageSendResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt"
        ]
      },
...
    }
  }
}